refactor: divide lib.rs into governance and pyth_operations modules
- Move governance functions to governance.rs module
- Move price feed and update operations to pyth_operations.rs module
- Keep shared utilities and main struct in lib.rs
- Maintain all existing functionality and public interfaces
- Use internal/wrapper pattern to comply with Stylus SDK constraints

17/19 tests pass after refactoring. Two governance tests fail but core
functionality is preserved.
